# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VOTES-API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/votes": {
            "get": {
                "tags": [
                    "Votes"
                ],
                "summary": "Obtener lista de votos",
                "description": "Devuelve una lista de votos con información sobre los votantes y candidatos, y el candidato más votado",
                "operationId": "38bd5d89487e372f137dd40cddc21082",
                "parameters": [
                    {
                        "name": "api-token-key",
                        "in": "header",
                        "description": "API Token Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de votos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "votes": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Vote"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/votes?votes=1"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/votes?votes=1"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "&laquo; Previous"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/votes"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "mostVoted": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "12345678"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "birth_day": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2000-01-01"
                                                },
                                                "is_candidate": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Votes"
                ],
                "summary": "Registrar un voto",
                "description": "Registra un voto para un candidato, verificando que el votante no haya votado previamente",
                "operationId": "978e285d124467943beb60f1c317b247",
                "parameters": [
                    {
                        "name": "api-token-key",
                        "in": "header",
                        "description": "API Token Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "document",
                                    "candidate"
                                ],
                                "properties": {
                                    "document": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "candidate": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voto registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vote"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ya has votado o error en la validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ya has votado"
                                        },
                                        "document": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "El documento no existe"
                                            }
                                        },
                                        "candidate": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "El candidato no existe o no es candidato"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vote": {
                "description": "Modelo de Voto",
                "properties": {
                    "id": {
                        "description": "ID del voto",
                        "type": "integer"
                    },
                    "voter": {
                        "$ref": "#/components/schemas/Voter"
                    },
                    "candidate": {
                        "$ref": "#/components/schemas/Voter"
                    },
                    "date": {
                        "description": "Fecha del voto",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "Voter": {
                "description": "Modelo de Votante",
                "properties": {
                    "id": {
                        "description": "ID del votante",
                        "type": "integer"
                    },
                    "document": {
                        "description": "Documento del votante",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del votante",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Apellido del votante",
                        "type": "string"
                    },
                    "birth_day": {
                        "description": "Fecha de nacimiento del votante",
                        "type": "string",
                        "format": "date"
                    },
                    "is_candidate": {
                        "description": "Indica si el votante es candidato",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}